Convert h.schemas into a package
In preparation for moving schemas from memex, this commit converts
h.schemas into a package. It also moves the "base" class CSRFSchema from
h.accounts.schemas into h.schemas.base, in preparation for moving the
various schemas defined in h.accounts into h.schemas.
