+++ h/schemas/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+"""Classes for validating data supplied to the application."""
+
+from h.schemas.base import ValidationError
+
+__all__ = ('ValidationError',)
