+++ h/schemas/base.py
@@ -1,15 +1,44 @@
 # -*- coding: utf-8 -*-
-"""Classes for validating data passed to views."""
+
+"""Shared functionality for schemas."""
 
 from __future__ import unicode_literals
 
 import copy
+
+import colander
+import deform
 import jsonschema
 from jsonschema.exceptions import best_match
+from pyramid.session import check_csrf_token
+
+
+@colander.deferred
+def deferred_csrf_token(node, kw):
+    request = kw.get('request')
+    return request.session.get_csrf_token()
 
 
 class ValidationError(Exception):
     pass
+
+
+class CSRFSchema(colander.Schema):
+    """
+    A CSRFSchema backward-compatible with the one from the hem module.
+
+    Unlike hem, this doesn't require that the csrf_token appear in the
+    serialized appstruct.
+    """
+
+    csrf_token = colander.SchemaNode(colander.String(),
+                                     widget=deform.widget.HiddenWidget(),
+                                     default=deferred_csrf_token,
+                                     missing=None)
+
+    def validator(self, form, value):
+        request = form.bindings['request']
+        check_csrf_token(request)
 
 
 class JSONSchema(object):
