--- tests/h/schemas_test.py
@@ -1,36 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from __future__ import unicode_literals
-
-import pytest
-
-from h import schemas
-
-
-class ExampleSchema(schemas.JSONSchema):
-    schema = {
-        b'$schema': b'http://json-schema.org/draft-04/schema#',
-        b'type': b'string',
-    }
-
-
-class TestJSONSchema(object):
-    def test_it_returns_data_when_valid(self):
-        data = "a string"
-
-        assert ExampleSchema().validate(data) == data
-
-    def test_it_raises_when_data_invalid(self):
-        data = 123  # not a string
-
-        with pytest.raises(schemas.ValidationError):
-            ExampleSchema().validate(data)
-
-    def test_it_sets_appropriate_error_message_when_data_invalid(self):
-        data = 123  # not a string
-
-        with pytest.raises(schemas.ValidationError) as e:
-            ExampleSchema().validate(data)
-
-        message = e.value.message
-        assert message.startswith("123 is not of type 'string'")
